NhanDangChuVietTay_mnist_streamlit/requirements.txt
tensorflow==2.10.0
numpy==1.23.5
opencv-python==4.7.0.72
streamlit==1.22.0
Pillow==9.4.0
streamlit-drawable-canvas==0.9.3 